refactor(connectivity): consolidate to Tier 1 single mod.rs
